requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
celery==5.3.4
redis==5.0.1
cryptography==41.0.7
Pillow==10.1.0
pandas==2.1.3
python-nmap==0.7.1
pyyaml==6.0.1
ansible-core==9.8.0
numpy==1.24.3
paramiko==3.4.0
pyvmomi==8.0.2.0.1
requests==2.31.0
humanfriendly==10.0
openpyxl==3.1.2
stevedore==5.2.0
python-dotenv==1.2.1